% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{abundance_interval_boundaries}
\alias{abundance_interval_boundaries}
\title{Table des intervalles d’abondance}
\format{
Un data frame avec les colonnes suivantes :
\describe{
  \item{code}{Code d’abondance (numérique)}
  \item{start}{Borne inférieure de l’intervalle d’abondance}
  \item{end}{Borne supérieure de l’intervalle d’abondance}
}
}
\source{
Grille d’abondance interne SIPV, adaptée du protocole InfoFlora.
}
\usage{
abundance_interval_boundaries
}
\description{
Cet objet fournit la correspondance entre les codes d’abondance et les intervalles numériques
utilisés pour estimer les effectifs d’observations.
Il peut être utilisé directement dans la fonction \code{create_clean_data()}.
}
\examples{
data(abundance_interval_boundaries)
head(abundance_interval_boundaries)

}
\keyword{datasets}
