% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{columns_map}
\alias{columns_map}
\title{Mapping des colonnes pour le format SIPV}
\format{
Une liste nommée avec les éléments suivants :
\describe{
  \item{obs_id}{Identifiant unique de la note d’observation (\code{NO_NOTE})}
  \item{taxon}{Nom scientifique complet de l’espèce (\code{NOM_COMPLET})}
  \item{observer}{Nom de l’observateur ou des observateurs (\code{OBSERVATEURS})}
  \item{day, month, year}{Date d’observation (\code{J_NOTE}, \code{M_NOTE}, \code{A_NOTE})}
  \item{canton, commune}{Localisation administrative (\code{CO_CANTON}, \code{NOM_COMMUNE})}
  \item{x, y}{Coordonnées (\code{X}, \code{Y})}
  \item{precision}{Précision de la localisation (\code{XY_PRECISION})}
  \item{xy_type}{Type de coordonnées (\code{XY_FORME})}
  \item{origine}{Origine ou statut d’introduction de l’espèce (\code{INTRODUIT})}
  \item{presence_absence}{Présence ou absence (\code{PRESENCE})}
  \item{unity}{Unité de comptage (\code{UNITE_COMPTAGE})}
  \item{abundance_code}{Catégorie d’abondance (\code{CAT_ABONDANCE})}
  \item{abundance_indication}{Valeur d’abondance observée (\code{ABONDANCE})}
  \item{effectif}{Effectif observé ou calculé (\code{EFFECTIF})}
}
}
\source{
Données issues du système d’information SIPV (Service d’Information sur les Plantes Vasculaires)
}
\usage{
columns_map
}
\description{
Cet objet contient la correspondance entre les noms de colonnes d’un jeu de données
brut (tel que fourni par le système SIPV) et ceux attendus par la fonction
\code{create_clean_data()}.
}
\details{
Chaque élément de la liste associe un nom standardisé (utilisé dans le package)
au nom réel de la colonne dans le jeu de données brut.
}
\keyword{datasets}
