% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_valid_data.R
\name{filter_valid_data}
\alias{filter_valid_data}
\title{Filter valid data from an InfoFlora-like dataset}
\usage{
filter_valid_data(
  data,
  presence_col = "presence_absence",
  abundance_indication_col = "abundance_indication",
  abundance_code_col = "abundance_code",
  unity_col = "unity",
  absence_values = c("N", "I", "P", "-"),
  colors = c(Usable = "#5D7052", `Potentially usable` = "#FCD581", Unusable = "#A7001E")
)
}
\arguments{
\item{data}{A dataframe}

\item{presence_col}{Name of the column containing presence/absence info (default: "presence_absence")}

\item{abundance_indication_col}{Name of the column with abundance indications (default: "abundance_indication")}

\item{abundance_code_col}{Name of the column with abundance codes (default: "abundance_code")}

\item{unity_col}{Name of the column with the unit info (default: "unity")}

\item{absence_values}{Vector of values considered as absence (default: c("N", "I", "P", "-"))}

\item{colors}{Named character vector of colors for the categories ('Usable', 'Potentially usable', 'Unusable')}
}
\value{
A **list** with:
\itemize{
 \item \code{data}: Filtered dataset (includes 'Usable' and 'Potentially usable' records).
 \item \code{percent_usable}: Percentage of fully usable records.
 \item \code{percent_potential}: Percentage of potentially usable records.
 \item \code{percent_unusable}: Percentage of unusable records.
 \item \code{plot}: A \code{ggplot2} object (pie chart) visualizing the breakdown.
}
}
\description{
Filter valid data from an InfoFlora-like dataset
}
\examples{
# Cas avec noms standards
filter_valid_data(df)

# Cas avec colonnes personnalisees
filter_valid_data(df,
                  presence_col = "presence",
                  abundance_indication_col = "abundance_text",
                  abundance_code_col = "ab_code",
                  unity_col = "unit")

}
