% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trim_trend.R
\name{plot_trim_trend}
\alias{plot_trim_trend}
\title{Plot TRIM model trend}
\usage{
plot_trim_trend(model)
}
\arguments{
\item{model}{A TRIM model object}
}
\value{
A ggplot2 plot showing abundance trend over time
}
\description{
Plot TRIM model trend
}
