% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resume_data.R
\name{resume_data}
\alias{resume_data}
\title{Resume data for a species with table and plots}
\usage{
resume_data(donneessp, colonnesite, colonneannee, colonneeffectif)
}
\arguments{
\item{donneessp}{Dataframe **prepared for TRIM** (\code{prepare_data_for_trim} output) for a given species.}

\item{colonnesite}{Character: column name containing site}

\item{colonneannee}{Character: column name containing year}

\item{colonneeffectif}{Character: column name containing abundance}
}
\value{
A **list** with three elements:
\itemize{
 \item \code{resume_table}: Data frame summarizing sites, years, and missing data percentages.
 \item \code{plot_points}: A \code{ggplot2} scatter plot of abundance (log10) over time.
 \item \code{plot_heatmap}: A \code{ggplot2} heatmap of abundance by site and year.
}
}
\description{
Generates a comprehensive summary of the time series data for a given species,
including statistics on missing data and visualization plots (log-transformed scatter plot and heatmap).
}
