% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/application_secure_trim.R
\name{safe_wald}
\alias{safe_wald}
\title{Safely run rtrim::wald}
\usage{
safe_wald(model)
}
\arguments{
\item{model}{TRIM model object}
}
\value{
The result of rtrim::wald or NULL if an error occurs
}
\description{
Safely run rtrim::wald
}
