% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{synonymes_template}
\alias{synonymes_template}
\title{Example synonym list template}
\format{
A list of character vectors.
}
\usage{
synonymes_template
}
\description{
This object provides an example list structure for defining taxon synonyms.
Each element is a character vector where the first name is the canonical (accepted)
name and the following names are its synonyms.
}
\examples{
synonymes <- synonymes_template
synonymes[[1]][1] <- "Aira caryophyllea L."  # canonical name
synonymes[[1]][2] <- "Aira caryophyllea subsp. caryophyllea"  # synonym
}
\keyword{datasets}
